Changed PmergeMe algorithm

+++ C09_Module/ex01/RPN.hpp
@@ -17,8 +17,8 @@ class RPN {
         RPN(const RPN& other);
         RPN& operator = (const RPN& other);
 		RPN(std::string input);
-		long firstOp(std::stack<std::string>& newStack);
-		long doOp(long num1, long num2, std::string operat);
+		float firstOp(std::stack<std::string>& newStack);
+		float doOp(float num1, float num2, std::string operat);
 		bool isNumber(const std::string token);
 		bool isOperator(const std::string token);
 		void validateOrder(std::stack<std::string>& newStack);
